[ROCM private] Change the default warp size to 64 for IR compatibility for existing MI GPUs.
CAVEAT: This commit is supposed to be a custom patch in ROCM/triton fork.
        Think twice before submitting this commit as PR to upstream.

MI Developers have collected a large set of IR files and use triton
command line tools for development extensively, where warp size == 64 is
assumed for AMD GPUs. Unfortunately the behavior of the compiler after the TTGIR pass
depends on the warp size property, and changing the default value will
make existing IRs unusable for MI GPUs.

This commit aims to preserve the old behavior when warp size is not specified in TTGIR.
For general Triton users it should have zero effects since warp size is
always set explicitly in compiler.py to match the target architecture.